Attempt to fix the javadoc error
@MatthewL246 I think this should fix the Javadoc generation. We just need to change
```yaml
- name: Generate documentation with Javadoc
        run:
          javadoc -d javadoc-output $(find . -name "*.java") -Xdoclint:none
          --ignore-source-errors
```
in javadoc.yml to something like
```yaml
- name: Generate documentation with Javadoc
        run:
          gradle docs
```
Could you test this on your fork? You'll need to reference a dependency from inside the code, for example make a documented method like:
```java
/**
* @param opMode some message
*/
public void method(LinearOpMode opMode) {}
```
That should force the generator to index a dependency, and if it works it should be fixed.